Quanta-bot/production/requirements.txt
streamlit
pandas
websockets
python-dotenv
aiosqlite
aiohttp
